+++ src/Compze.Internals.SystemCE/IOCE/IBinaryFile.cs
@@ -1,6 +1,6 @@
 namespace Compze.Internals.SystemCE.IOCE;
 
-public interface IBinaryFile
+public interface IBinaryFile : IDisposable
 {
    byte[] ReadAllBytes();
    void WriteAllBytes(byte[] bytes);